pubsub_listener/pyproject.toml
[project]
name = "pubsub-listener"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-cloud-pubsub>=2.31.0",
    "requests>=2.32.4",
]
